fix: make CompactionPlan pickle-able and import-able
